requirements.txt
discord.py==2.3.2
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
